RMS-Worker-dev-sheetal/RMS-Worker-dev-sheetal/requirements.txt
python-dotenv
pydantic
redis
jsons

pymongo
motor

PyMuPDF
pdf2image
python-docx
docx2pdf

openai
langchain-core
langchain
langchain-community
langchain-openai

pandas
bs4

starlette
asyncpg
openai-agents